Refactored the KeyBuilder into a base class without Unicode support (light weight) and a UnicodeKeyBuilder (heavier weight and not always required).
Modified the BigdataClient to cache the metadata index locally and moved to a static partitioning scheme for scale-out indices as a first step.  I am still working through the client API for transparent access to the scale-out index partitions and then I will update the rdf application to use the client API.  I have not yet done the asynchronous overflow handling or the quad store migration.


git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@209 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/scaleup/ResourceState.java
@@ -54,14 +54,17 @@ import com.bigdata.btree.IndexSegment;
  */
 public enum ResourceState {
 
-    New("New",0),
-    Live("Live",1),
-    Dead("Dead",2);
+    New("New",(short)0),
+    Live("Live",(short)1),
+    Dead("Dead",(short)2);
     
     final private String name;
-    final private int id;
+    final private short id;
     
-    ResourceState(String name,int id) {this.name = name;this.id = id;}
+    ResourceState(String name, short id) {
+        this.name = name;
+        this.id = id;
+    }
     
     public String toString() {return name;}
